# backend/src/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v2/users": {
      "get": {
        "description": "Retrieve all users (admin or authenticated users). Supports cursor-based pagination.",
        "operationId": "UsersController_find_all",
        "parameters": [
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Sort order (default: desc)",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sort_by",
            "required": false,
            "in": "query",
            "description": "Sort by field (default: created_at)",
            "schema": {
              "enum": [
                "created_at",
                "email",
                "name"
              ],
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Base64 encoded cursor for cursor-based pagination",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (1-100, default: 25)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api-key": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "description": "Create a new user account (admin or authenticated users)",
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User data",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "email": "user@example.com",
                  "name": "John Doe",
                  "avatar_url": "https://..."
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api-key": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a user",
        "tags": [
          "users"
        ]
      }
    },
    "/v2/users/{id}": {
      "get": {
        "description": "Retrieve a specific user's profile",
        "operationId": "UsersController_find_by_id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "api-key": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "users"
        ]
      },
      "put": {
        "description": "Update a user account. Requires admin or ownership.",
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Updated user data",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "email": "new@example.com",
                  "name": "Jane Doe"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Can only update own account"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "api-key": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a user",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "description": "Remove a user account. Requires admin or ownership.",
        "operationId": "UsersController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Can only delete own account"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "api-key": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete a user",
        "tags": [
          "users"
        ]
      }
    },
...
    "/v2/reactions": {
      "get": {
        "description": "Retrieve reactions with filtering, sorting, and pagination (offset or cursor-based)",
        "operationId": "ReactionsController_find_all",
        "parameters": [
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Sort order (default: desc)",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sort_by",
            "required": false,
            "in": "query",
            "description": "Sort by field (default: created_at)",
            "schema": {
              "enum": [
                "created_at",
                "reaction_type"
              ],
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "required": false,
            "in": "query",
            "description": "Filter by user UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reaction_type",
            "required": false,
            "in": "query",
            "description": "Filter by reaction type",
            "schema": {
              "enum": [
                "like",
                "love",
                "insight",
                "question",
                "flag"
              ],
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": false,
            "in": "query",
            "description": "Filter by comment UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Base64 encoded cursor for cursor-based pagination",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (1-100, default: 25)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of reactions"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all reactions",
        "tags": [
          "reactions"
        ]
      },
      "post": {
        "description": "Add a new reaction to a comment",
        "operationId": "ReactionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Reaction data",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "comment_id": "uuid",
                  "user_id": "uuid",
                  "reaction_type": "like"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reaction created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflict - Reaction already exists"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a reaction",
        "tags": [
          "reactions"
        ]
      }
    },
    "/v2/reactions/{id}": {
      "put": {
        "description": "Update an existing reaction",
        "operationId": "ReactionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Reaction UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Updated reaction data",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "reaction_type": "love"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reaction updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Reaction not found"
          },
          "409": {
            "description": "Conflict - Reaction already exists with this type"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a reaction",
        "tags": [
          "reactions"
        ]
      },
      "delete": {
        "description": "Remove a reaction",
        "operationId": "ReactionsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Reaction UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reaction deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Reaction not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete a reaction",
        "tags": [
          "reactions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ohara API",
    "description": "REST API for Ohara App developed por EnSer management system with AI agent integration",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Admin API key for server-to-server authentication"
      }
    },
...
  }
}